refactor(algebras): rename alegebra.js the plural algebras.js
