+++ public/models/fixtures/client-project.js
@@ -1,7 +1,7 @@
 import fixture from 'can-fixture';
 import json from './client-projects.json';
 //import { algebra } from '../client-project';
-import { _idAlgebra as algebra } from '../algebra';
+import { _idAlgebra as algebra } from '../algebras';
 
 var store = fixture.store(json, algebra);
