+++ public/models/fixtures/contribution-months.js
@@ -2,7 +2,7 @@ import fixture from 'can-fixture';
 import json from './contribution-months.json';
 // Can't import algebra from the model since it will initiate connection.
 //import { algebra } from '../contribution-month';
-import { _idAlgebra as algebra } from '../algebra';
+import { _idAlgebra as algebra } from '../algebras';
 
 export var store = fixture.store(json, algebra);
