+++ public/models/fixtures/contributor.js
@@ -1,6 +1,6 @@
 import fixture from 'can-fixture';
 //import Contributor from '../contributor';
-import { _idAlgebra as algebra } from '../algebra';
+import { _idAlgebra as algebra } from '../algebras';
 
 var store = fixture.store([{
     _id: "1-JustinMeyer",
