+++ public/models/fixtures/os-project.js
@@ -1,7 +1,7 @@
 import fixture from 'can-fixture';
 import json from './os-projects.json';
 //import { algebra } from '../os-project';
-import { _idAlgebra as algebra } from '../algebra';
+import { _idAlgebra as algebra } from '../algebras';
 
 export var store = fixture.store(json, algebra);
