+++ public/models/signup.js
@@ -3,7 +3,7 @@ import DefineList from 'can-define/list/';
 import set from 'can-set';
 import feathersClient from './feathers';
 import superModel from '../lib/super-model';
-import { _idAlgebra as signupAlgebra } from './algebra';
+import { _idAlgebra as signupAlgebra } from './algebras';
 
 export const Signup = DefineMap.extend('Signup', {
   _id: '*',
